styled-components를 활용한 Theme 설정 * 전역에 사용할 스타일을 위한 GlobalStyle 추가   - styled-reset이나 styled-noramlize를 사용하여 브라우저마다 설정 된 기본 스타일을 초기화   - 전역에서 사용하고자 하는 스타일을 별도로 추가
