+++ package.json
@@ -11,7 +11,8 @@
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
     "styled-components": "^5.3.3",
-    "web-vitals": "^2.1.4"
+    "styled-normalize": "^8.0.7",
+    "styled-reset": "^4.3.4"
   },
   "scripts": {
     "start": "react-scripts start",
